Added option for another fans
Added an option to include other types of fans, such as a 12V DC fan. Note that the fan's effectiveness will be effected by it's normal current/voltage and the amount of batteries you use to drive your motor driver with. A 12V DC fan runs just find on 7.5 Volts (0.45 Amps)